api_gateway/requirements.txt
Django==3.0.2
djangorestframework==3.10.3
django-webpack-loader==0.6.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.2.1
requests==2.22.0
uuid==1.30